backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic",
    "bcrypt>=5.0.0",
    "celery",
    "dotenv>=0.9.9",
    "eventlet",
    "fastapi",
    "fastapi-mail>=1.6.2",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]",
    "pwdlib[argon2]>=0.3.0",
    "pydantic[email]>=2.12.5",
    "python-dotenv",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart",
    "qrcode>=8.2",
    "redis",
    "reportlab>=4.4.10",
    "sqlalchemy[asyncio]",
    "uvicorn[standard]",
    "websockets>=16.0",
]
[tool.uv]
required-environments = ["sys_platform == 'win32' and platform_machine == 'AMD64'"]
